Simplify API interface
I seriously over-engineered the API for the first cut. There is NO need for an API user
to have their own DSFAppearanceManager instance, so now the entire interface
works through the static implementation. Also, there's no reason to publicly advertise
a generic change api, so it was removed.

I've also removed a lot of clutter and apis that should never have been public.