Fixed issue where sometimes the UserDefaults is not updated with dark mode
The UserDefaults doesn't always seem to be set when the appearance changes. For example, setting the
mode via the Shortcuts app doesn't set it, but via System Preferences does.

Now falls back to the default NSApp.effectiveAppearance setting.
Note that this might not always be accurate as the app may have explicitly set the effective appearance.