Dealing with AutoplayImages crashes with Xcode 16 beta
AXAnimatedImagesEnabled/NSNotification.Name.AXAnimatedImagesEnabledDidChange were introduced
in macOS 14, but is DEPRECATED for macOS 15.

Compiling apps that target macOS 14 using Xcode 16 beta create apps that RUNTIME CRASH whenever
they attempt to access either of these symbols. It appears that some library is no longer automatically
linked when using Xcode 15.

This is a (hacky) work around for this.  Situation is :-

* If we compile using Xcode 15 or lower and run macOS 14, we use the old symbols which work
* If we compile using Xcode 16 and run on macOS 14, we drop back to using the REDUCE MOTION setting to work around this issue
* If we compile using Xcode 16 and run on macOS 15, we use the new settings.

Annoying but only affects a single case (and the autoplay images setting is probably not used much anyway)